m2-task2: update tests.
