+++ src/components/app/app.test.js
@@ -1,11 +1,12 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
 import App from './app.jsx';
+import filmList from '../../mock/filmList.js';
 
 it(`App correctly renders after relaunch`, () => {
   const tree = renderer
     .create(<App
-      filmList={[]}
+      filmList={filmList}
       onClick={jest.fn()}
     />)
     .toJSON();
